refactor: Update test imports to use index.ts
